Default the Buzzer and LED to enabled
on startup enable the busser and the LED. I made this change to my reader after noticing that the restore state dose does not work most of the time. Defaulting to enabled comes from feedback from noticing the tag users scanning the tag more than ones thinking that the tag was not scanned.